requirements.txt
beautifulsoup4==4.13.5
bidict==0.23.1
jsonlines==4.0.0
lightning==2.5.5
matplotlib==3.10.0
openai==1.102.0
pandas==2.3.1
pandarallel==1.6.5
playwright
pytest-playwright
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.5
seaborn==0.13.2
scikit-learn==1.7.1
spotipy==2.25.1
streamlit==1.50.0
swifter==1.3.5
wandb==0.21.3
--extra-index-url https://download.pytorch.org/whl/cu124
# GPU-enabled torch wheels come from here
torch
torchvision
torchaudio